move console_script tests back into integrated files
