+++ tests/test_aiohttp.py
@@ -1017,6 +1017,44 @@ async def test_aiohttp_using_cli(event_loop, aiohttp_server, monkeypatch, capsys
 
 
 @pytest.mark.asyncio
+@pytest.mark.script_launch_mode("subprocess")
+async def test_aiohttp_using_cli_ep(
+    event_loop, aiohttp_server, monkeypatch, script_runner, run_sync_test
+):
+    from aiohttp import web
+
+    async def handler(request):
+        return web.Response(text=query1_server_answer, content_type="application/json")
+
+    app = web.Application()
+    app.router.add_route("POST", "/", handler)
+    server = await aiohttp_server(app)
+
+    url = str(server.make_url("/"))
+
+    def test_code():
+
+        monkeypatch.setattr("sys.stdin", io.StringIO(query1_str))
+
+        ret = script_runner.run(
+            "gql-cli", url, "--verbose", stdin=io.StringIO(query1_str)
+        )
+
+        assert ret.success
+
+        # Check that the result has been printed on stdout
+        captured_out = str(ret.stdout).strip()
+
+        expected_answer = json.loads(query1_server_answer_data)
+        print(f"Captured: {captured_out}")
+        received_answer = json.loads(captured_out)
+
+        assert received_answer == expected_answer
+
+    await run_sync_test(event_loop, server, test_code)
+
+
+@pytest.mark.asyncio
 async def test_aiohttp_using_cli_invalid_param(
     event_loop, aiohttp_server, monkeypatch, capsys
 ):
